New test case. Other changes.
New test case: tc_add_course_pages.rb

New method added to Utilities module (though specific to OAE, so it's
in the app_functions.rb file.

Did some code housekeeping in Page Elements to improve readability of
the classes and modules. This will be an on-going project as I continue
to add comments to classes and methods.